Refactor interceptor to multiple implementations to allow supporting new platforms
// TODO:
- Expose a public API for specifying a custom interceptor
- Make sure remote mode is working
- Test ReactNative
- Test NativeScript
