+++ lib/api/index.js
@@ -1,5 +1,5 @@
 import Emitter from 'api/emitter';
-import Interceptor from 'api/interceptor';
+import WebInterceptor from 'api/interceptors/web-interceptor';
 import { PersistentStorage } from 'api/storage';
 import Mocks from 'api/mocks';
 import Groups from 'api/groups';
@@ -16,7 +16,8 @@ let interceptor;
 
 // only init the interceptor if we're running inside a browser
 if (!get(global, 'isReactNative') && !get(global, 'isNativeScript')) {
-  interceptor = new Interceptor('window');
+  interceptor = new WebInterceptor();
+  interceptor.init();
 }
 
 class PublicAPI {
@@ -133,7 +134,7 @@ class PublicAPI {
   import(data = '{}', options) {
     return new Promise((resolve, reject) => {
       if (interceptor) {
-        return interceptor.communicator.waitForInit()
+        return interceptor.waitForInit()
           .then(() => {
             try {
               const dataTree = JSON.parse(data);
