+++ lib/api/remote.js
@@ -1,9 +1,9 @@
-import Interceptor from 'api/interceptor';
+import RemoteInterceptor from 'api/interceptors/remote-interceptor';
 import API from 'api';
 
 function connect({ hostname = 'localhost', port = 5000 } = {}) {
   API.setMode('remote');
-  new Interceptor('remote', { hostname, port });
+  new RemoteInterceptor({ hostname, port });
 
   return API;
 }
