+++ lib/api/requests.js
@@ -1,5 +1,7 @@
 import { Request } from 'api/models/request';
 import find from 'lodash/find';
+import EVENTS from 'api/constants/events';
+import Emitter from 'api/emitter';
 
 const capturedRequests = [];
 
@@ -17,18 +19,22 @@ export default class Requests {
     });
 
     capturedRequests.push(capturedRequest);
+
+    Emitter.emit(EVENTS.REQUEST_CAPTURED, { requestId: request.id });
   }
 
   static setResponse(id, response, startTime) {
     const request = find(capturedRequests, { id });
-
+    
     if (request) {
       request.response = {
         status: response.status || 200,
         delay: response.status === 0 ? 0 : (Date.now() - startTime),
         headers: response.headers || {},
         body: response.data || ''
-      }
+      };
+
+      Emitter.emit(EVENTS.RESPONSE_RECEIVED, { requestId: id });
     }
   }
