+++ lib/api/worker.js
@@ -1,7 +1,7 @@
 import EVENTS from 'api/constants/events';
-import Interceptor from 'api/interceptor';
+import WebWorkerInterceptor from 'api/interceptors/web-worker-interceptor';
 
-const interceptor = new Interceptor('worker');
+const interceptor = new WebWorkerInterceptor();
 
 self.addEventListener('message', (event) => {
   if (event.data.type === EVENTS.MIMIC_TURN_ON) {
